+++ xc3_lib/src/mxmd.rs
@@ -455,7 +455,7 @@ pub struct Material {
 pub struct MaterialFlags {
     pub unk1: bool,
     pub unk2: bool,
-    /// Enables alpha testing from a texture when `true`.
+    /// Enables alpha testing from a texture in a prepass when `true`.
     pub alpha_mask: bool,
     /// Samples `texture.x` from a dedicated mask texture when `true`.
     /// Otherwise, the alpha channel is used.